Moved text_run mod to it's own folder, fixed update_lines_text_run() method, added utility methods to TextRun struct. Added some docs and formatted changed files.
